requirements.txt
# ==========================================
# Core
# ==========================================
numpy==1.26.4
pandas==2.2.2

# ==========================================
# Deep Learning (CPU)
# ==========================================
tensorflow==2.16.1   # Works well with Python 3.11

# ==========================================
# Image Processing
# ==========================================
Pillow==10.3.0

# ==========================================
# Similarity Search
# ==========================================
faiss-cpu==1.7.4

# ==========================================
# Web App
# ==========================================
streamlit==1.35.0

# ==========================================
# Optional (if needed later)
# ==========================================
scikit-learn==1.4.2